pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "korean-anki-pipeline"
version = "0.1.0"
description = "Local workflow to turn Korean lesson material into reviewed Anki cards."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "openai>=1.109.0",
  "pydantic>=2.7.0",
  "python-dotenv>=1.0.1",
  "requests>=2.32.0",
]

[project.scripts]
korean-anki = "korean_anki.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
